Base Script
Making a Base Script for the towers to make the implementation of new tower types easier
