+++ Assets/Scripts/Tower/TowerTypes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dce6fc76f4626034693c921524b73819
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
