+++ Assets/Scripts/Tower/TowerTypes/SimpleTower.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 16c1371fe24976e40a801fde202c0086
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
